requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
click==8.0.4
fastapi==0.89.1
frozenlist==1.3.3
h11==0.12.0
idna==3.4
multidict==6.0.4
pip==22.3.1
pydantic==1.10.4
python-dotenv==0.21.1
setuptools==66.1.1
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.16.0
wheel==0.38.4
yarl==1.8.2
